lazy load swap form
